requirements.txt
Flask==2.3.3
python-telegram-bot==13.15
googletrans==4.0.0-rc1
